refactored service layer for first few entities

+++ Maven/ACME/src/main/java/ENSF614Group1/ACME/Service/BankService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ENSF614Group1.ACME.Repository.*;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import ENSF614Group1.ACME.Model.*;
 
@@ -16,24 +17,43 @@ import java.util.Optional;
 public class BankService {
 
 	@Autowired
-    private BankRepository bankRepository;
+	private BankRepository bankRepository;
+	
+	@Transactional
+	public Bank createBank(Bank bank) {
+		return bankRepository.save(bank);
+	}
+	
+	public List<Bank> getAllBanks(){
+		return bankRepository.findAll();
+	}
+	
+	public Bank getBankById(Long id){
+		Optional<Bank> bank = bankRepository.findById(id);
+		if(bank.isEmpty()) {
+			throw new EntityNotFoundException("Bank does not exist.");
+		}
+		return bank.get();
+	}
 
-    @Transactional
-    public Bank createBank(Bank bank) {
-        return bankRepository.save(bank);
-    }
-
-    public Optional<Bank> getBankById(Long id) {
-        return bankRepository.findById(id);
-    }
-
-    public List<Bank> getAllBanks() {
-        return bankRepository.findAll();
-    }
-
-    @Transactional
-    public void deleteBank(Long id) {
-        bankRepository.deleteById(id);
-    }
+	
+	@Transactional
+	public Bank updateBank(Long id, Bank bankDetails) {
+		Optional<Bank> bank = bankRepository.findById(id);
+		if (bank.isEmpty()) {
+			throw new EntityNotFoundException("Bank does not exist.");
+		}
+		Bank s = bank.get();
+		s.setTitle(bankDetails.getTitle());
+		return bankRepository.save(s);
+	}
+	
+	@Transactional
+	public void deleteBank(Long id) {
+		if (!bankRepository.existsById(id)) {
+			throw new EntityNotFoundException("Bank does not exist.");
+		}
+		bankRepository.deleteById(id);
+	}
     
 }
