+++ Maven/ACME/src/main/java/ENSF614Group1/ACME/Service/CreditCardService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ENSF614Group1.ACME.Repository.*;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import ENSF614Group1.ACME.Model.*;
 
@@ -16,24 +17,49 @@ import java.util.Optional;
 public class CreditCardService {
 
 	@Autowired
-    private CreditCardRepository creditCardRepository;
+	private CreditCardRepository creditCardRepository;
+	
+	@Transactional
+	public CreditCard createCreditCard(CreditCard creditCard) {
+		return creditCardRepository.save(creditCard);
+	}
+	
+	public List<CreditCard> getAllCreditCards(){
+		return creditCardRepository.findAll();
+	}
+	
+	public CreditCard getCreditCardById(Long id){
+		Optional<CreditCard> creditCard = creditCardRepository.findById(id);
+		if(creditCard.isEmpty()) {
+			throw new EntityNotFoundException("CreditCard does not exist.");
+		}
+		return creditCard.get();
+	}
 
-    @Transactional
-    public CreditCard createCreditCard(CreditCard creditCard) {
-        return creditCardRepository.save(creditCard);
-    }
+	
+	@Transactional
+	public CreditCard updateCreditCard(Long id, CreditCard creditCardDetails) {
+		Optional<CreditCard> creditCard = creditCardRepository.findById(id);
+		if (creditCard.isEmpty()) {
+			throw new EntityNotFoundException("CreditCard does not exist.");
+		}
+		CreditCard s = creditCard.get();
+		
+		s.setFirstName(creditCardDetails.getFirstName());
+		s.setLastName(creditCardDetails.getLastName());
+		s.setCardNumber(creditCardDetails.getCardNumber());
+		s.setExpiry(creditCardDetails.getExpiry());
+		s.setBank(creditCardDetails.getBank());
 
-    public Optional<CreditCard> getCreditCardById(Long id) {
-        return creditCardRepository.findById(id);
-    }
-
-    public List<CreditCard> getAllCreditCards() {
-        return creditCardRepository.findAll();
-    }
-
-    @Transactional
-    public void deleteCreditCard(Long id) {
-        creditCardRepository.deleteById(id);
-    }
+		return creditCardRepository.save(s);
+	}
+	
+	@Transactional
+	public void deleteCreditCard(Long id) {
+		if (!creditCardRepository.existsById(id)) {
+			throw new EntityNotFoundException("CreditCard does not exist.");
+		}
+		creditCardRepository.deleteById(id);
+	}
     
 }
