+++ Maven/ACME/src/main/java/ENSF614Group1/ACME/Service/RegisteredUserService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ENSF614Group1.ACME.Repository.*;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import ENSF614Group1.ACME.Model.*;
 
@@ -16,29 +17,48 @@ import java.util.Optional;
 public class RegisteredUserService {
 
 	@Autowired
-    private RegisteredUserRepository registeredUserRepository;
-    
+	private RegisteredUserRepository registeredUserRepository;
+	
     @PersistenceContext
     private EntityManager entityManager;
+	
+	@Transactional
+	public RegisteredUser createRegisteredUser(RegisteredUser registeredUser) {
+		return registeredUserRepository.save(registeredUser);
+	}
+	
+	public List<RegisteredUser> getAllRegisteredUsers(){
+		return registeredUserRepository.findAll();
+	}
+	
+	public RegisteredUser getRegisteredUserById(Long id){
+		Optional<RegisteredUser> registeredUser = registeredUserRepository.findById(id);
+		if(registeredUser.isEmpty()) {
+			throw new EntityNotFoundException("RegisteredUser does not exist.");
+		}
+		return registeredUser.get();
+	}
 
-
-    @Transactional
-    public RegisteredUser createRegisteredUser(RegisteredUser user) {
-        return registeredUserRepository.save(user);
-    }
-
-    public Optional<RegisteredUser> getRegisteredUserById(Long id) {
-        return registeredUserRepository.findById(id);
-    }
-
-    public List<RegisteredUser> getAllRegisteredUsers() {
-        return registeredUserRepository.findAll();
-    }
-
-    @Transactional
-    public void deleteRegisteredUser(Long id) {
-        registeredUserRepository.deleteById(id);
-    }
+	
+	@Transactional
+	public RegisteredUser updateRegisteredUser(Long id, RegisteredUser registeredUserDetails) {
+		Optional<RegisteredUser> registeredUser = registeredUserRepository.findById(id);
+		if (registeredUser.isEmpty()) {
+			throw new EntityNotFoundException("RegisteredUser does not exist.");
+		}
+		RegisteredUser s = registeredUser.get();
+		s.setMembershipExpires(registeredUserDetails.getMembershipExpires());
+		s.setCreditCard(registeredUserDetails.getCreditCard());
+		return registeredUserRepository.save(s);
+	}
+	
+	@Transactional
+	public void deleteRegisteredUser(Long id) {
+		if (!registeredUserRepository.existsById(id)) {
+			throw new EntityNotFoundException("RegisteredUser does not exist.");
+		}
+		registeredUserRepository.deleteById(id);
+	}
     
     @Transactional
     public RegisteredUser register(User user, LocalDateTime membershipExpires, CreditCard creditCard) {
